Refactor auth logic
Save public user data in local storage. Repopulate user if it's null
and user data exists in local storage. If a user tries to do something
and gets a 401, that means they don't have a valid auth token (in
cookie form), so remove user data from state and local storage.