Use proper app structure for front end
For all features, have folders for components, api, util, and types, and
export items as needed from an index file. Components should then import
items from the index file rather than from individual files. The
exception to this is components importing other components from within
the same feature directory, which can import directly.

The auth feature is left as-is because that needs to be refactored
anyway.