Implement first draft of IDB logic for userPuzzleData
The app seems to work where it worked previously, which is for logged in
users. There is still a lot of placeholder functionality, particularly
for the logic to compare the data retrieved from IndexedDB vs the
server. Also, logic for syncing UUIDs after a collision hasn't been
implemented yet, and most of the logic for writing to IDB still hasn't
been implemented.