Allow OpenAI service to recover from words not found
When iterating through the word hints, the initial functionality was to
let an ActiveRecord::RecordNotFound error bubble up to the top function,
killing the seed hints loop. After running it a few times, however, it's
clear that the AI will sometimes misspell words, so it's better to just
skip over those rather than stop the whole process.